# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Visileaf API",
    "description": "Visileaf API Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/enhancement/enhance": {
      "post": {
        "tags": [
          "enhancement"
        ],
        "summary": "Enhance Image",
        "description": "Enhance image with specified methods.\n\n- **image**: Image file (JPEG, PNG, etc.)\n- **params_json**: Enhancement parameters as JSON string\n\nExample JSON:\n```json\n{\n    \"use_gamma\": true,\n    \"gamma\": 0.5,\n    \"use_msr\": true,\n    \"msr_sigmas\": [15, 80, 250],\n    \"use_clahe\": true,\n    \"clahe_clip\": 2.5,\n    \"clahe_tile_size\": [8, 8],\n    \"use_ssr\": false,\n    \"ssr_sigma\": 80,\n    \"use_sharpen\": false,\n    \"sharpen_method\": \"unsharp\",\n    \"sharpen_strength\": 1.0,\n    \"sharpen_kernel_size\": 5,\n    \"order\": [\"gamma\", \"msr\", \"clahe\"],\n    \"use_lowlight_lime\": true,\n    \"use_lowlight_dual\": false,\n    \"lowlight_gamma\": 0.6,\n    \"lowlight_lambda\": 0.15,\n    \"lowlight_sigma\": 3.0,\n    \"lowlight_bc\": 1.0,\n    \"lowlight_bs\": 1.0,\n    \"lowlight_be\": 1.0\n}\n```\n\nReturns:\n    Enhanced image as downloadable JPEG file",
        "operationId": "enhance_image_api_enhancement_enhance_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enhance_image_api_enhancement_enhance_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/enhancement/enhance-json": {
      "post": {
        "tags": [
          "enhancement"
        ],
        "summary": "Enhance Image Json",
        "description": "Enhance image with specified methods (alternative endpoint with JSON body).\n\n- **image**: Image file (JPEG, PNG, etc.) - sent as FormData\n- **params**: Enhancement parameters - sent as JSON body\n\nNote: This endpoint uses multipart/form-data for the image and JSON body for parameters.\nSome clients may have issues with this combination. Use /enhance endpoint for better compatibility.\n\nReturns:\n    Enhanced image as downloadable JPEG file",
        "operationId": "enhance_image_json_api_enhancement_enhance_json_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enhance_image_json_api_enhancement_enhance_json_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/enhancement/enhance-with-dcp": {
      "post": {
        "tags": [
          "enhancement"
        ],
        "summary": "Enhance Image With Dcp",
        "description": "Enhance image using Dark Channel Prior (DCP) based low-light enhancement.\n\n- **image**: Image file (JPEG, PNG, etc.)\n- **params_json** (optional): Enhancement parameters as JSON string.\n\nNot: `params_json` şu an algoritma içinde kullanılmıyor, ileride DCP parametrelerini\n(ör. patch_size, omega, t_min vb.) kontrol etmek için ayrılmıştır.",
        "operationId": "enhance_image_with_dcp_api_enhancement_enhance_with_dcp_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enhance_image_with_dcp_api_enhancement_enhance_with_dcp_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/enhancement/dcp-guided-filter": {
      "post": {
        "tags": [
          "enhancement"
        ],
        "summary": "Enhance Image With Dcp Guided",
        "description": "Enhance image using Dark Channel Prior (DCP) + Guided Filter based advanced low-light enhancement.\n\n- **image**: Image file (JPEG, PNG, etc.)\n- **params_json** (optional): Enhancement parameters as JSON string.\n\nNot: `params_json` şu an algoritma içinde kullanılmıyor, ileride patch_size, radius,\neps vb. guided filter / DCP parametrelerini kontrol etmek için ayrılmıştır.",
        "operationId": "enhance_image_with_dcp_guided_api_enhancement_dcp_guided_filter_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enhance_image_with_dcp_guided_api_enhancement_dcp_guided_filter_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/enhancement/with-deep-learning": {
      "post": {
        "tags": [
          "enhancement"
        ],
        "summary": "Enhance Image With Deep Learning",
        "description": "Derin öğrenme tabanlı hazır modelleri kullanarak enhancement.\n\n- **image**: İşlenecek görsel dosyası (JPEG, PNG, vb.)\n- **params_json**: Form-data alanı, içinde `model_name` beklenir. Örn:\n    `{\"model_name\": \"enlightengan\"}`\n\nNot: EnlightenGAN servisine entegre, sonuç DB'ye kaydedilir.",
        "operationId": "enhance_image_with_deep_learning_api_enhancement_with_deep_learning_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enhance_image_with_deep_learning_api_enhancement_with_deep_learning_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/enhancement/results": {
      "get": {
        "tags": [
          "enhancement"
        ],
        "summary": "List Enhancement Results",
        "description": "List recent enhancement results for the current user.\n\nReturns:\n    List of records containing input/output image info and params.",
        "operationId": "list_enhancement_results_api_enhancement_results_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_enhance_image_api_enhancement_enhance_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "Image file to enhance"
          },
          "params_json": {
            "type": "string",
            "title": "Params Json",
            "description": "Enhancement parameters as JSON string"
          }
        },
        "type": "object",
        "required": [
          "image",
          "params_json"
        ],
        "title": "Body_enhance_image_api_enhancement_enhance_post"
      },
      "Body_enhance_image_json_api_enhancement_enhance_json_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "Image file to enhance"
          },
          "params": {
            "$ref": "#/components/schemas/EnhancementParams",
            "description": "Enhancement parameters"
          }
        },
        "type": "object",
        "required": [
          "image",
          "params"
        ],
        "title": "Body_enhance_image_json_api_enhancement_enhance_json_post"
      },
      "Body_enhance_image_with_dcp_api_enhancement_enhance_with_dcp_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image",
            "description": "Image file to enhance with Dark Channel Prior (DCP)"
          },
          "params_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params Json",
            "description": "Optional enhancement parameters as JSON string (currently not used by DCP algorithm, reserved for future use)"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_enhance_image_with_dcp_api_enhancement_enhance_with_dcp_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}